CTAA-1383: Adapt remaining reporting flows, refactor flow
- Self Testing (probably sick)
- Revocation (probably sick)
- Revocation (sick)
